+++ example_project/assets/svelte/HelloWorld.svelte
@@ -1,1 +1,1 @@
-Hello World
+<span data-testid="hello-world-content">Hello World</span>